Cleaning : use static_cast where possible, compile with -Wall

+++ src/glGribColorbar.cc
@@ -95,13 +95,13 @@ void glGrib::Colorbar::render (const glm::mat4 & MVP, const glGrib::Palette & p,
         {
           if (opts.levels.values.size () > 0)
             {
-              for (int i = 0; i < opts.levels.values.size (); i++)
+              for (size_t i = 0; i < opts.levels.values.size (); i++)
                 if ((min <= opts.levels.values[i]) && (opts.levels.values[i] <= max))
                   values.push_back (opts.levels.values[i]);
             }
           else if (values_pal.size () > 0)
             {
-              for (int i = 0; i < values_pal.size (); i++)
+              for (size_t i = 0; i < values_pal.size (); i++)
                 if ((min <= values_pal[i]) && (values_pal[i] <= max))
                   values.push_back (values_pal[i]);
             }
@@ -112,7 +112,7 @@ void glGrib::Colorbar::render (const glm::mat4 & MVP, const glGrib::Palette & p,
                 values.push_back (min + d * i);
             }
           
-          for (int i = 0; i < values.size (); i++)
+          for (size_t i = 0; i < values.size (); i++)
             {
               float val = values[i];
               char tmp[32];
@@ -131,7 +131,7 @@ void glGrib::Colorbar::render (const glm::mat4 & MVP, const glGrib::Palette & p,
         }
       else
         {
-          for (int i = 0; i < values_pal.size (); i++)
+          for (size_t i = 0; i < values_pal.size (); i++)
             {
               float val = values_pal[i];
               char tmp[32];
@@ -146,7 +146,7 @@ void glGrib::Colorbar::render (const glm::mat4 & MVP, const glGrib::Palette & p,
             }
 
 	  rank2rgba[0] = 0;
-          for (int i = 0; i < values_pal.size () - 1; i++)
+          for (size_t i = 0; i < values_pal.size () - 1; i++)
             {
               int j1 = 1 + (255 * (i + 0)) / (values_pal.size () - 1);
               int j2 = 1 + (255 * (i + 1)) / (values_pal.size () - 1);
